requirements.txt
PyYAML==5.3.1
pyagme==2.0.1
pygame_gui==0.5.7 
